Update for node tree id provider. Missing a few last minute changes that needed variable casting.
git-svn-id: https://svn.apache.org/repos/asf/incubator/vxquery/trunk/vxquery@1465891 13f79535-47bb-0310-9956-ffa450edef68

+++ vxquery-core/src/main/java/org/apache/vxquery/metadata/VXQueryCollectionOperatorDescriptor.java
@@ -62,7 +62,7 @@ public class VXQueryCollectionOperatorDescriptor extends AbstractSingleActivityO
         final ArrayBackedValueStorage abvsFileNode = new ArrayBackedValueStorage();
         final int partitionId = ctx.getTaskAttemptId().getTaskId().getPartition();
         // TODO Add the data source scan id.
-        final ITreeNodeIdProvider nodeIdProvider = new TreeNodeIdProvider(partitionId);
+        final ITreeNodeIdProvider nodeIdProvider = new TreeNodeIdProvider((short) partitionId);
 
         return new AbstractUnaryInputUnaryOutputOperatorNodePushable() {
             @Override
